Added the Easy TileData system!
It allows you to easily sync your TileEntities simply by adding an Annotation to the fields to sync! It will do everything all by itself and even supports custom objects.
Also improved the recipe system. It allows ItemStacks which are null now, too. (They're not actually null when checking because of a small bug. They are "converted" into a stack of bedrock)
Also swapped all the mappings to Guava Tables for the sake of performance and simplicity.